+++ lib/crisp_view.dart
@@ -1,9 +1,22 @@
+import 'package:crisp/helpers.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'models/main.dart';
+
+const CRISP_BASE_URL = 'https://go.crisp.chat';
+
+String crispEmbedUrl(String websiteId, String locale) {
+  String url = CRISP_BASE_URL + '/chat/embed/?website_id=$websiteId';
+
+  if (notNull(locale)) url += '&locale=$locale';
+
+  return url;
+}
+
+final crisp = CrispMain();
 
 class CrispView extends StatefulWidget {
   final Widget loadingWidget;
@@ -21,8 +34,8 @@ class _CrispViewState extends State<CrispView> {
   handleAppLifecycleState() {
     SystemChannels.lifecycle.setMessageHandler((msg) {
       if (msg == "AppLifecycleState.resumed") {
-        flutterWebViewPlugin.reloadUrl(
-            'https://go.crisp.chat/chat/embed/?website_id=${crisp.websiteId}');
+        flutterWebViewPlugin
+            .reloadUrl(crispEmbedUrl(crisp.websiteId, crisp.locale));
       }
       return null;
     });
@@ -52,7 +65,7 @@ class _CrispViewState extends State<CrispView> {
       }
 
       if (state.type == WebViewState.shouldStart) {
-        if (state.url.contains('https://go.crisp.chat')) return;
+        if (state.url.contains(CRISP_BASE_URL)) return;
 
         print("navigating to...${state.url}");
         if (state.url.startsWith("mailto") ||
@@ -81,7 +94,7 @@ class _CrispViewState extends State<CrispView> {
   @override
   Widget build(BuildContext context) {
     return WebviewScaffold(
-      url: 'https://go.crisp.chat/chat/embed/?website_id=${crisp.websiteId}',
+      url: crispEmbedUrl(crisp.websiteId, crisp.locale),
       mediaPlaybackRequiresUserGesture: false,
       appBar: widget.appBar,
       withZoom: true,